magit-section-goto-successor: use magit-section-goto for sibling hunks
When going to a sibling hunk of the previously current hunk which just
disappeared, then use `magit-section-goto' to go there.  In this case
we want to run `magit-section-movement-hook' in case the user has added
`magit-hunk-maybe-recenter' to that.  Other hook functions should not
cause an issues as they either have no affect when going to a hunk
section, or in the case of `magit-hunk-set-window-start' have a
desirable effect.

Closes #2683.